requirements.txt
fastapi
uvicorn
gunicorn
openai
python-dotenv
pinecone
PyPDF2
python-docx
requests